+++ src/OGLRender.h
@@ -73,7 +73,7 @@ public:
     virtual void TexCoord(TLITVERTEX &vtxInfo);
 
     void DrawSimple2DTexture(float x0, float y0, float x1, float y1, float u0, float v0, float u1, float v1, COLOR dif, COLOR spe, float z, float rhw);
-    void DrawSimpleRect(LONG nX0, LONG nY0, LONG nX1, LONG nY1, uint32 dwColor, float depth, float rhw);
+    void DrawSimpleRect(int nX0, int nY0, int nX1, int nY1, uint32 dwColor, float depth, float rhw);
     void InitCombinerBlenderForSimpleRectDraw(uint32 tile=0);
     void DrawSpriteR_Render();
     void DrawObjBGCopy(uObjBg &info);
